fix(iter-2 #89): drop renderer-side strips; close .canonicalize() leaks in resolve_sidecar
Pre-flight: rg -n '\.canonicalize\(\)' src-tauri/src
  before: 7 hits (3 prod in core/paths.rs:116/126/133 + 4 in mod tests)
  after:  0 hits

Task 1  verified hypothesis was wrong. WatcherState::is_path_or_parent_allowed
  AND is_path_allowed already canonicalize input via canonicalize_no_verbatim
  (a dunce::canonicalize wrapper). Empirically verified that dunce DOES
  resolve Windows 8.3 short names (RUNNER~1 -> runneradmin) — so the
  starts_with check succeeds in both directions. Added a Windows-only
  regression test path_form_test::is_path_or_parent_allowed_accepts_8dot3_short_name_input
  that pins the existing-file, deleted-file, and is_path_allowed paths
  against 8.3 short-name input. The native E2E failure must have a
  different root cause; deferring further investigation.

Task 2  Delete renderer-side stripVerbatimPrefix defensive calls in
  useFolderTree (ghost dedupe) and folder-tree.ts
  (pathStartsWithRootCrossPlatform). Producer-side fix in
  core::paths::canonicalize_no_verbatim is the chokepoint; renderer guards
  hide upstream regressions. Helper kept in src/lib/paths.ts for the
  Zustand v1 migrate callback. Dropped the corresponding 'verbatim form'
  unit/component test cases (issue-#89 cross-form ghost dedupe and
  'Other files' filter; folder-tree.ts cross-form prefix tests).

Task 3  Replace 3 .canonicalize() method-form leaks in
  core::paths::resolve_sidecar with canonicalize_no_verbatim (CLI stdout
  + future-proofing if any Tauri command wraps it). Updated the unit
  tests to assert against canonicalize_no_verbatim form (proves the fix
  took effect  pre-fix tests asserted std::fs verbatim form).

Task 4  File-size budget compliance:
  - src-tauri/src/core/paths.rs 408 -> 157 LOC; tests extracted to
    core/paths_tests.rs and included via #[path = ...] mod tests.
  - src/store/index.ts 529 -> 488 LOC; v1 migrate body extracted to
    src/store/migrations/v1-strip-verbatim.ts exporting
    migrateV1StripVerbatim.

Task 5  Docs:
  - docs/security.md rule 11 references canonicalize_no_verbatim with
    issue-#89 rationale; added rule 11a forbidding .canonicalize() in
    IPC-producing / persisting / frontend-comparing code.
  - docs/security.md rule 22 line range refreshed to watcher.rs:295-324.
  - docs/architecture.md rule 15 documents persist version + migrate fn,
    pointing at store/migrations/v1-strip-verbatim.ts.

Task 6  Replaced 'as unknown as ReturnType<NonNullable<...>>' double
  cast in store/index.ts with a single PersistedUI type defined inside
  migrations/v1-strip-verbatim.ts (used via 's as PersistedUI').

Validation (all green): npm run lint (0 errors); tsc --noEmit (clean);
  vitest run (1415 passed); cargo test --no-fail-fast (309+23+61+7+5+5
  passed across all binaries); playwright browser e2e (130 passed).
  Native E2E cannot run on this dev host; is_path_or_parent_allowed_accepts_8dot3_short_name_input
  is the pre-merge proof for the Windows 8.3 short-name path.

Addresses expert-panel iter 1: lean-BLOCKING, arch-MAJOR x3, docs-MAJOR x3.
