+++ docs/security.md
@@ -27,7 +27,8 @@ Canonical for threat-model and safety rules. Cite violations as "violates rule N
 ### Launch-args & CLI handling
 9. The `get_launch_args` handler drains a pending-args queue (`PendingArgsState`) so each batch of launch args is consumed exactly once. (`commands/launch.rs:15-55`.)
 10. Single-instance emits `args-received` only when `get_webview_window("main")` is `Some`. (`lib.rs:95-102`.)
-11. CLI argument parsing canonicalizes every path via `std::fs::canonicalize` and silently drops paths that fail. (`commands/launch.rs:68-125`; sidecar resolution under `core/paths.rs::resolve_sidecar` adds stricter folder-root containment.)
+11. CLI argument parsing canonicalizes every path via `core::paths::canonicalize_no_verbatim` (which strips Windows `\\?\` verbatim prefixes — see issue #89) and silently drops paths that fail. (`commands/launch.rs:68-125`; sidecar resolution under `core/paths.rs::resolve_sidecar` adds stricter folder-root containment.)
+11a. Every Rust→TS path crosses the IPC boundary in non-verbatim form. All canonicalization in commands, watcher, scanner, and CLI goes through `core::paths::canonicalize_no_verbatim`. `std::fs::canonicalize` and `Path::canonicalize` method-form are forbidden in code that produces an IPC payload, persists a path, or compares against a frontend-supplied path — they leak the Windows `\\?\` verbatim prefix and desynchronise from the bare-form paths the frontend already holds (`read_dir` output, dialog results, persisted `root` / tabs / `recentItems`). Producer-side fix only — renderers must not paper over a regression with defensive strips. Issue #89.
 
 ### Markdown rendering safety
 12. Markdown rendering uses `rehype-raw` (to admit inline HTML for GitHub-style `<details>`/`<sub>`/`<kbd>` etc.) ONLY when paired with `rehype-sanitize` configured by the schema in `src/components/viewers/markdown/sanitizeSchema.ts`. The schema is the canonical XSS boundary: `script`/`iframe`/`object`/`embed`/`form` are absent from `tagNames` (and so dropped); `on*` event handler attributes are absent from per-tag and `*` attribute lists (and so stripped). `style` is allow-listed on `span` and `math` only as a narrow KaTeX exception (KaTeX emits inline `style` attributes for math layout); a custom `rehype-katex-style` preprocessor (`src/components/viewers/markdown/rehype-katex-style.ts`) walks the tree and STRIPS `style` from any `span`/`math` whose className does not start with `katex` AND whose nearest ancestor is not katex-classed, so raw HTML cannot smuggle styles through the KaTeX hole. Plugin order in `MarkdownViewer.tsx` MUST be `rehypeRaw` → `rehypeFootnotePrefix` → `rehypeKatex` (lazy) → `rehypeKatexStyle` → `rehypeSanitize(sanitizeSchema)` → downstream plugins so user HTML cannot piggy-back through later transforms. Footnote ids/hrefs are kept in sync via the default `clobberPrefix: "user-content-"`; `rehypeFootnotePrefix` strips a redundant prefix that `remark-gfm` itself sometimes emits, so sanitize re-adds exactly one prefix.
@@ -45,7 +46,7 @@ Canonical for threat-model and safety rules. Cite violations as "violates rule N
 
 ### Watcher integrity
 21. The watcher watches parent directories (not individual files) to survive atomic-rename saves, and emits only for paths on the current watch list. (`watcher.rs:146-169, 80-102`.) Debounce window: rule 4 in [`docs/performance.md`](performance.md).
-22. Watcher bookkeeping stores both canonical and raw paths so deleted files (which cannot canonicalize) still match. (`watcher.rs:184-197`.)
+22. Watcher bookkeeping stores both canonical and raw paths so deleted files (which cannot canonicalize) still match. (`watcher.rs:295-324`.)
 23. Closing a tab evicts that path from `lastSaveByPath` so stale timestamps cannot suppress a later event. (`store/index.ts:156-159`.)
 
 ### Logging & crash capture
